requirements.txt
# Guardrails Before the Fall - Experimental Dependencies
# Author: Clarence H. Cannon, IV
# Install: pip install -r requirements.txt

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ML/AI
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
anthropic>=0.18.0

# Evaluation & Benchmarking
datasets>=2.14.0
evaluate>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment Tracking
wandb>=0.15.0
tqdm>=4.65.0

# Data Processing
jsonlines>=3.1.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.25.0